api1/pyproject.toml
[tool.poetry]
name = "1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.18"
aiofiles = "23.1.0"
markdown = "3.4.3"
ruff = "0.0.275"
SQLAlchemy = "^2.0.21"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
urllib3 = "^2.0.6"
alembic = "^1.12.0"
click = "^8.1.7"
starlette-admin = "^0.11.2"
itsdangerous = "^2.1.2"
sqlalchemy-file = "^0.6.0"
fasteners = "^0.19"
Pillow = "^10.0.1"
SQLAlchemy-Utils = "^0.41.1"
pydantic = {extras = ["email"], version = "^2.4.2"}
pandas = "^2.1.1"
black = "^23.9.1"
pydantic-settings = "^2.0.3"
pyodbc = "^5.0.0"
asyncodbc = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
